monitor: use espflash configuration
Using the default() configuration, if 'y' is answered to the question:
"Remember this serial port for future use?" the program panic because
the `save_path` variable of the espflash Config is not initialized to
a valid directory to store a persistent configuration.

Using the load() function allows to skip the questions on next use.